Update prepare-host.sh
I don't know. This doesn't feel right. The package exists in Jammy, Noble and Bookworm in both amd64 and arm64 variants so should be good...though it is probably not needed in all constellations. 
Still it is part of the toolchain and not generic packages. If this gets merged as it is this should be revisited.