ti: configs: Update K3 kernel configs (#9019)
A couple of these configs have gotten out of sync with is making it
difficult for the automation scripting to properly re-write these
configs correctly. Mostly due to the "edge" kernel being updated
to the latest v6.18 kernel, but also due to the RT config not being
updated in some time. Manually update these configs and check for
correctness.

Signed-off-by: Andrew Davis <afd@ti.com>